// src/app/interfaces/user.ts
export interface User {
    displayName?: string;
    phoneNumber?: string;
    email?: string;
    country?: string;
    address?: string;
    balance?: number;
    balanceUsdt?: number;
    admin?: boolean;
    uid?: string;
    photoURL?: string;
    provider?: string;
    createDate?: number;
    role?: string;
    holdTime?: number;
    holdTimeUntil?: number;
    tyc?: boolean;
}